+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_firewall_policy_idps_signatures_filter_values_operations.py
@@ -25,7 +25,7 @@ class TestNetworkManagementFirewallPolicyIdpsSignaturesFilterValuesOperations(Az
             resource_group_name=resource_group.name,
             firewall_policy_name="str",
             parameters={"filterName": "str"},
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself