+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_public_ip_prefixes_operations_async.py
@@ -26,7 +26,7 @@ class TestNetworkManagementPublicIPPrefixesOperationsAsync(AzureMgmtRecordedTest
             await self.client.public_ip_prefixes.begin_delete(
                 resource_group_name=resource_group.name,
                 public_ip_prefix_name="str",
-                api_version="2024-05-01",
+                api_version="2024-07-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -39,7 +39,7 @@ class TestNetworkManagementPublicIPPrefixesOperationsAsync(AzureMgmtRecordedTest
         response = await self.client.public_ip_prefixes.get(
             resource_group_name=resource_group.name,
             public_ip_prefix_name="str",
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself
@@ -70,9 +70,12 @@ class TestNetworkManagementPublicIPPrefixesOperationsAsync(AzureMgmtRecordedTest
                         "name": "str",
                         "provisioningState": "str",
                         "publicIpAddresses": [{"id": "str"}],
+                        "publicIpAddressesV6": [{"id": "str"}],
                         "publicIpPrefixes": [{"id": "str"}],
+                        "publicIpPrefixesV6": [{"id": "str"}],
                         "resourceGuid": "str",
                         "sku": {"name": "str"},
+                        "sourceVirtualNetwork": {"id": "str"},
                         "subnets": [{"id": "str"}],
                         "tags": {"str": "str"},
                         "type": "str",
@@ -88,7 +91,7 @@ class TestNetworkManagementPublicIPPrefixesOperationsAsync(AzureMgmtRecordedTest
                     "type": "str",
                     "zones": ["str"],
                 },
-                api_version="2024-05-01",
+                api_version="2024-07-01",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -102,7 +105,7 @@ class TestNetworkManagementPublicIPPrefixesOperationsAsync(AzureMgmtRecordedTest
             resource_group_name=resource_group.name,
             public_ip_prefix_name="str",
             parameters={"tags": {"str": "str"}},
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself
@@ -112,7 +115,7 @@ class TestNetworkManagementPublicIPPrefixesOperationsAsync(AzureMgmtRecordedTest
     @recorded_by_proxy_async
     async def test_public_ip_prefixes_list_all(self, resource_group):
         response = self.client.public_ip_prefixes.list_all(
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself
@@ -123,7 +126,7 @@ class TestNetworkManagementPublicIPPrefixesOperationsAsync(AzureMgmtRecordedTest
     async def test_public_ip_prefixes_list(self, resource_group):
         response = self.client.public_ip_prefixes.list(
             resource_group_name=resource_group.name,
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself