+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_routing_rules_operations_async.py
@@ -27,7 +27,7 @@ class TestNetworkManagementRoutingRulesOperationsAsync(AzureMgmtRecordedTestCase
             network_manager_name="str",
             configuration_name="str",
             rule_collection_name="str",
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself
@@ -42,7 +42,7 @@ class TestNetworkManagementRoutingRulesOperationsAsync(AzureMgmtRecordedTestCase
             configuration_name="str",
             rule_collection_name="str",
             rule_name="str",
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself
@@ -76,7 +76,7 @@ class TestNetworkManagementRoutingRulesOperationsAsync(AzureMgmtRecordedTestCase
                 },
                 "type": "str",
             },
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself
@@ -92,7 +92,7 @@ class TestNetworkManagementRoutingRulesOperationsAsync(AzureMgmtRecordedTestCase
                 configuration_name="str",
                 rule_collection_name="str",
                 rule_name="str",
-                api_version="2024-05-01",
+                api_version="2024-07-01",
             )
         ).result()  # call '.result()' to poll until service return final result